xburst: zx10: adjust device for modified U-Boot
WARNING
This breaks compatibility of produced images with original U-Boot. The
factory.bin image is still usable in original web interface but the old
U-Boot won't be able to boot it. If the image has size lower than 10240
KiB, original U-Boot will fallback to backup system but if it's bigger,
You'll need to use serial console to recover.

The source of this modification can be obtained at:
https://github.com/tmn505/u-boot/tree/zx10
while automatically compilled binaries are at:
https://tmn505.github.io/u-boot
Use the latest available.

To know what exactly was changed, inspect the git log. The most
important changes are:
- increased read size of kernel to 8MiB
- remove kernel integrity check and backup fallback
- add posibility to boot from µSD card

Installation
1. Gain telnet access to original firmware command line as mentioned in
   "xburst: initial support for ZoomGo Media Stick" commit, in
   "Aditional information" section.
2. Copy modified U-boot binary to FAT32 formatted µSD card and insert
   the card to the slot.
3. While connected through telnet execute this command (adjust
   accordingly):
    mtd -r write /tmp/mnt/SD-Card-1/new_uboot_binary_filename bootloader
4. Device will reboot after successful writing with new U-Boot in place.

- or -

1. Having soldered serial console, push new U-Boot using serial
   connection. Remember, this is a Little Endian MIPS system.
2. Execute following commands (adjust accordingly):
    sfcnor erase 0x0 0x30000
    sfcnor write 0x0 0xsize_of_new_uboot 0xmemory_address_of_copied_new_uboot
    reset
3. Device will reboot with new U-Boot in place.

Following on the change, now we don't have to squeeze kernel to 2MiB
image, thus remove the small_flash optimisation.

Signed-off-by: Tomasz Maciej Nowak <tmn505@gmail.com>