kernel: modules: enable UART variant of Broadcom bluetooth
Bunch of SBC integrate SDIO Broadcom BT+WiFi module, but while WiFi uses
SDIO lines, the BT part usually is connected to UART lines, instead of
SDIO, like most would asume, so package module responsible for it to
unlock this feature. Caveat is, this driver needs firmware which isn't
contained in linux-firmware repo, so one needs to find it.

Signed-off-by: Tomasz Maciej Nowak <tmn505@gmail.com>